trying to deconstruct server.js for testing - lots of problems
