+++ server/package.json
@@ -22,7 +22,7 @@
     "pg": "^8.14.1",
     "puppeteer": "^24.6.1",
     "winston": "^3.17.0",
-    "ws": "^8.18.1"
+    "ws": "^8.18.2"
   },
   "devDependencies": {
     "@babel/core": "^7.26.10",
